requirements.txt
asgiref==3.6.0
boto3==1.26.142
botocore==1.29.142
Django==4.1.7
django-cors-headers==3.14.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.4.0
psycopg2==2.9.6
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.16
whitenoise==6.4.0
